// packages/xla-sdk-core/src/generated/XLAPrepaymentFactory_v010.ts
export declare namespace XLARSCPrepaymentFactory {
  export type RSCCreateDataStruct = {
    controller: PromiseOrValue<string>;
    distributors: PromiseOrValue<string>[];
    isImmutableController: PromiseOrValue<boolean>;
    isAutoNativeCurrencyDistribution: PromiseOrValue<boolean>;
    minAutoDistributeAmount: PromiseOrValue<BigNumberish>;
    investor: PromiseOrValue<string>;
    investedAmount: PromiseOrValue<BigNumberish>;
    interestRate: PromiseOrValue<BigNumberish>;
    residualInterestRate: PromiseOrValue<BigNumberish>;
    initialRecipients: PromiseOrValue<string>[];
    percentages: PromiseOrValue<BigNumberish>[];
    supportedErc20addresses: PromiseOrValue<string>[];
    erc20PriceFeeds: PromiseOrValue<string>[];
    creationId: PromiseOrValue<BytesLike>;
  };

  export type RSCCreateDataStructOutput = [
    string,
    string[],
    boolean,
    boolean,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    string[],
    BigNumber[],
    string[],
    string[],
    string
  ] & {
    controller: string;
    distributors: string[];
    isImmutableController: boolean;
    isAutoNativeCurrencyDistribution: boolean;
    minAutoDistributeAmount: BigNumber;
    investor: string;
    investedAmount: BigNumber;
    interestRate: BigNumber;
    residualInterestRate: BigNumber;
    initialRecipients: string[];
    percentages: BigNumber[];
    supportedErc20addresses: string[];
    erc20PriceFeeds: string[];
    creationId: string;
  };
}